Refactor model out of api     -UI tweaks:         -better date formatting         -correct page header
