+++ tests/api_tests/tokens/test_views.py
@@ -3,8 +3,7 @@ import mock
 
 from nose.tools import *  # flake8: noqa
 
-from api.tokens.models import ApiOAuth2PersonalToken
-from website.models import User
+from website.models import User, ApiOAuth2PersonalToken
 from website.util import api_v2_url
 from website.util import sanitize
