+++ website/static/js/apiPersonalToken.js
@@ -18,6 +18,8 @@ var koHelpers = require('./koHelpers');  // URL validators etc
 var $osf = require('./osfHelpers');
 var oop = require('js/oop');
 var language = require('js/osfLanguage');
+var moment = require('moment');
+
 
 /*
  *  Store the data related to a single API Personal Token
@@ -38,7 +40,8 @@ var TokenData = oop.defclass({
 
         // Other fields. Owner and client ID should never change within this view.
         this.id = data.id;
-        this.lastUsed = attributes.date_last_used;
+        this.lastUsed = moment.utc(attributes.date_last_used).toString().split(' GMT')[0] + ' UTC';
+        debugger;
 
         this.owner = attributes.user_id;
         this.webDetailUrl = data.links ? data.links.html : undefined;
